pyproject.toml
[project]
name = "multi-model-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "langchain[google-genai]>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "langchain-text-splitters>=0.3.11",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "langchain-google-genai>=2.1.10",
    "langchain-chroma>=0.2.5",
]
